moved database to common
